requirements.txt
IMDbPY ~= 2020.9.25
discord.py ~= 1.5.0
python-dotenv ~= 0.14.0
discord~=1.0.1
Pillow~=7.2.0
PyNaCL~=1.4.0